+++ CotEditor/Sources/AppDelegate.swift
@@ -103,10 +103,8 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         super.awakeFromNib()
         
         // append the current version number to "What’s New" menu item
-        let shortVersion = Bundle.main.shortVersion
-        let shortVersionRange = shortVersion.range(of: "^[0-9]++\\.[0-9]++", options: .regularExpression)!
         self.whatsNewMenuItem?.title = String(format: "What’s New in CotEditor %@".localized,
-                                              String(shortVersion[shortVersionRange]))
+                                              Bundle.main.minorVersion)
         
         // sync menus with setting list updates
         EncodingManager.shared.$encodings
@@ -148,6 +146,13 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     
     // MARK: Application Delegate
     
+    /// Enable secure restorable state
+    func applicationSupportsSecureRestorableState(_ app: NSApplication) -> Bool {
+        
+        true
+    }
+    
+    
     #if SPARKLE
     /// setup Sparkle framework
     func applicationWillFinishLaunching(_ notification: Notification) {
@@ -160,14 +165,29 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     /// just after application did launch
     func applicationDidFinishLaunching(_ notification: Notification) {
         
-        // setup KeyBindingManager
         MenuKeyBindingManager.shared.applyKeyBindingsToMainMenu()
         
-        // register Services
         NSApp.servicesProvider = ServicesProvider()
-        
-        // setup touchbar
+        NSHelpManager.shared.registerBooks(in: .main)
         NSTouchBar.isAutomaticCustomizeTouchBarMenuItemEnabled = true
+        
+        // show notification panel for line ending migration on CotEditor 4.2.0
+        if let lastVersion = UserDefaults.standard[.lastVersion].flatMap(Int.init),
+           lastVersion < 494  // earlier than CotEditor 4.2.0
+        {
+            var migrationOptions: LineEndingMigrationOptions = []
+            if ReplacementManager.shared.needsLineEndingMigration() {
+                migrationOptions.insert(.replacement)
+            }
+            if SyntaxManager.shared.needsLineEndingMigration() {
+                migrationOptions.insert(.syntax)
+            }
+            if ScriptManager.shared.hasScripts {
+                migrationOptions.insert(.script)
+            }
+            
+            LineEndingMigrationPanel(options: migrationOptions).makeKeyAndOrderFront(nil)
+        }
     }